requirements.txt
absl-py==1.3.0
aioice==0.7.6
aiortc==1.3.2
albumentations==0.5.2
altair==4.2.0
appdirs==1.4.4
asttokens==2.2.0
attrs==22.1.0
audioread==3.0.0
autopep8==2.0.0
av==9.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blinker==1.5
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
chainer==7.8.1
chainer-openpose==0.0.1
chainercv==0.13.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
confection==0.0.3
contourpy==1.0.6
cryptography==38.0.4
customtkinter==5.0.1
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
darkdetect==0.7.1
debugpy==1.6.4
decorator==5.1.1
deepspeech==0.9.3
deta
dnspython==2.2.1
dominate==2.7.0
engineering-notation==0.7.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
extendedos==0.1.6
face-alignment==1.3.5
fcn==6.4.20
ffmpeg-python==0.2.0
ffpyplayer==4.3.5
filelock==3.8.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
future==0.18.2
gdown==4.5.4
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
greenlet==2.0.1
grpcio==1.51.1
gTTS==2.3.0
h5py==3.7.0
idna==3.4
imageio==2.22.4
imgaug==0.4.0
importlib-metadata==5.1.0
importlib-resources==5.10.0
install==1.3.5
ipykernel==6.17.1
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.7
jupyter_core==5.1.0
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.7.0
llvmlite==0.31.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
mutagen==1.46.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==2.8.8
nltk==3.7
numba==0.48.0
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openai==0.25.0
opencv-python==4.4.0.40
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
parso==0.8.3
pathy==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
pooch==1.6.0
preshed==3.0.8
prompt-toolkit==3.0.33
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0
pydub==0.25.1
pyee==9.0.4
Pygments==2.13.0
pylibsrtp==0.7.1
Pympler==1.0.1
pynormalize==0.1.4
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
pysptk==0.2.0
python-dateutil==2.8.2
python-vlc==3.0.18121
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyworld==0.3.2
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.3.1
Resemblyzer==0.1.1.dev0
rich==12.6.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
Shapely==1.8.5.post1
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soundfile==0.11.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.44
srsly==2.4.5
stack-data==0.6.2
streamlit==1.10.0
streamlit-server-state==0.15.0
streamlit-webrtc==0.43.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorvision==0.1.dev2
thinc==8.1.5
threadpoolctl==3.1.0
tifffile==2022.10.10
tk==0.1.0
tk-tools==0.16.0
tkvideoplayer==2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
torch-tb-profiler==0.4.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
typer==0.7.0
types-pytz==2022.6.0.1
typing==3.7.4.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wasabi==0.10.1
watchdog==2.2.0
wcwidth==0.2.5
webrtcvad==2.0.10
Werkzeug==2.2.2
zipp==3.11.0
